requirements.txt
fire==0.4.0
mysql-connector-python==8.0.30
protobuf==3.20.1
six==1.16.0
termcolor==1.1.0
